Move core.itanium_mangler into core.utils
